requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.4
openpyxl>=3.1.0
streamlit>=1.28.0
